+++ cyberdrop_dl/constants.py
@@ -25,6 +25,7 @@ MISSING: Any = Sentinel("MISSING")
 HttpMethod = Literal["GET", "POST", "PUT", "DELETE", "OPTIONS", "HEAD", "TRACE", "PATCH", "QUERY"]
 
 MAIN_LOG_FILE: ContextVar[Path] = ContextVar("MAIN_LOG_FILE")
+USE_RETRY_PATH: ContextVar[bool] = ContextVar("USE_RETRY_PATH", default=False)
 
 DEFAULT_PARAMETER = Parameter(
     negative_iterable=[],