+++ CHANGELOG.md
@@ -29,6 +29,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support premium URLs (direct links)(GoFile)
 - FileditchAlbums support
 
+### Changed
+
+- Enable the `--force-original-path` option on retry commands
+- Ask the user if they want to use the original download path after choosing `retry failed downloads` (main menu)
+
 ### Fixed
 
 - Pagination of some profiles stopping early (Twitter/x.com)